requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
django-timestampedmodel
cryptography
